+++ src/main/java/com/marowak/service/PortfolioItemService.java
@@ -5,5 +5,5 @@ import com.marowak.response.portfolio.PortfolioItemResponse;
 import java.util.List;
 
 public interface PortfolioItemService {
-    List<PortfolioItemResponse> get(String ticker, Long sliceTypeId);
+    List<PortfolioItemResponse> get(String ticker, int sliceTypeId);
 }
